requirements.txt
numpy==1.18.1
torchgeometry==0.1.2
joblib==0.14.1
opencv-python==4.5.5.64
trimesh==3.6.30
pyrender==0.1.39
smplx==0.1.28
tqdm==4.46.0
loguru==0.5.1
scipy==1.4.1
pyYAML==5.3.1
yacs==0.1.8
matplotlib==3.3.0
scikit-image==0.16.2
pillow==7.1.2
pytorch-lightning==1.5.1
filterpy==1.4.5
flatten-dict==0.3.0
chumpy==0.70
setuptools==47.3.1
ninja==1.10
ipython==8.1.1
pudb
comet-ml
numba==0.55.1
torchshow
jpeg4py==0.1.4
albumentations
git+https://github.com/mkocabas/yolov3-pytorch.git
git+https://github.com/mkocabas/multi-person-tracker.git
git+https://github.com/kinshuk-h/pytube
